+++ core/acp-core/src/main/kotlin/pers/acp/core/dbcon/DbType.kt
@@ -6,7 +6,12 @@ import pers.acp.core.exceptions.EnumValueUndefinedException
  * @author zhang by 10/07/2019
  * @since JDK 11
  */
-enum class DbType(private val namePrefix: String, private val nameSuffix: String, val instanceName: String, val isNoSQL: Boolean) {
+enum class DbType(
+    private val namePrefix: String,
+    private val nameSuffix: String,
+    val instanceName: String,
+    val isNoSQL: Boolean
+) {
 
     MySQL("`", "`", "MySQLInstance", false),
     Oracle("\"", "\"", "OracleInstance", false),
@@ -27,15 +32,15 @@ enum class DbType(private val namePrefix: String, private val nameSuffix: String
 
         init {
             for (type in values()) {
-                nameMap[type.name.toUpperCase()] = type
+                nameMap[type.name.uppercase()] = type
             }
         }
 
         @JvmStatic
         @Throws(EnumValueUndefinedException::class)
         fun getEnum(name: String): DbType {
-            if (nameMap.containsKey(name.toLowerCase())) {
-                return nameMap.getValue(name.toLowerCase())
+            if (nameMap.containsKey(name.lowercase())) {
+                return nameMap.getValue(name.lowercase())
             }
             throw EnumValueUndefinedException(DbType::class.java, name)
         }